+++ include/nihilus/common/monolithic_dispatcher.hpp
@@ -29,7 +29,7 @@ namespace nihilus {
 
 	template<model_config config, device_type dev_type, single_input core_type> struct kernel_dispatcher
 		: public kernel_traits<core_type::type, core_type::krn_type, core_type, typename core_type::output_type, typename core_type::input_type01::output_type> {
-		NIHILUS_FORCE_INLINE static void impl(core_type& params, size_t thread_index, size_t thread_count) {
+		NIHILUS_FORCE_INLINE static void impl(core_type& params, uint64_t thread_index, uint64_t thread_count) {
 			kernel_dispatcher_impl<cpu_arch_index, core_type::krn_type, typename core_type::transform_type, core_type, typename core_type::output_type,
 				typename core_type::input_type01::output_type>::impl(thread_index, thread_count, params, get_adjacent_value<config, core_type::type, 0>::impl(params));
 		}
@@ -38,7 +38,7 @@ namespace nihilus {
 	template<model_config config, device_type dev_type, double_input core_type> struct kernel_dispatcher<config, dev_type, core_type>
 		: public kernel_traits<core_type::type, core_type::krn_type, core_type, typename core_type::output_type, typename core_type::input_type01::output_type,
 			  typename core_type::input_type02::output_type> {
-		NIHILUS_FORCE_INLINE static void impl(core_type& params, size_t thread_index, size_t thread_count) {
+		NIHILUS_FORCE_INLINE static void impl(core_type& params, uint64_t thread_index, uint64_t thread_count) {
 			kernel_dispatcher_impl<cpu_arch_index, core_type::krn_type, typename core_type::transform_type, core_type, typename core_type::output_type,
 				typename core_type::input_type01::output_type, typename core_type::input_type02::output_type>::impl(thread_index, thread_count, params,
 				get_adjacent_value<config, core_type::type, 0>::impl(params), get_adjacent_value<config, core_type::type, 1>::impl(params));
@@ -48,7 +48,7 @@ namespace nihilus {
 	template<model_config config, device_type dev_type, triple_input core_type> struct kernel_dispatcher<config, dev_type, core_type>
 		: public kernel_traits<core_type::type, core_type::krn_type, core_type, typename core_type::output_type, typename core_type::input_type01::output_type,
 			  typename core_type::input_type02::output_type, typename core_type::input_type03::output_type> {
-		NIHILUS_FORCE_INLINE static void impl(core_type& params, size_t thread_index, size_t thread_count) {
+		NIHILUS_FORCE_INLINE static void impl(core_type& params, uint64_t thread_index, uint64_t thread_count) {
 			kernel_dispatcher_impl<cpu_arch_index, core_type::krn_type, typename core_type::transform_type, core_type, typename core_type::output_type,
 				typename core_type::input_type01::output_type, typename core_type::input_type02::output_type, typename core_type::input_type03::output_type>::impl(thread_index,
 				thread_count, params, get_adjacent_value<config, core_type::type, 0>::impl(params), get_adjacent_value<config, core_type::type, 1>::impl(params),
